fix: use bfloat16 instead of float16 to prevent NaN in Gemma logits
float16 overflows in attention computation (max 65504), causing
inf/nan probabilities during sampling. bfloat16 has wider dynamic
range and works on T4 via PyTorch software emulation.

+++ src/selene/core/med_logic.py
@@ -59,8 +59,8 @@ def _get_model():
                 _processor = AutoProcessor.from_pretrained(model_id, token=token)
 
                 if torch.cuda.is_available():
-                    dtype = torch.float16  # T4 supports fp16 natively
-                    device_map = {"": 0}   # entire model on cuda:0
+                    dtype = torch.bfloat16  # bf16 avoids fp16 overflow in Gemma attention
+                    device_map = {"": 0}    # entire model on cuda:0
                 else:
                     dtype = torch.float32
                     device_map = {"": "cpu"}
